+++ Cesium3DTilesContent/include/Cesium3DTilesContent/GltfConverterResult.h
@@ -2,8 +2,8 @@
 
 #include "Library.h"
 
-#include <Cesium3DTilesSelection/ErrorList.h>
 #include <CesiumGltf/Model.h>
+#include <CesiumUtility/ErrorList.h>
 
 #include <optional>
 
@@ -15,7 +15,7 @@ namespace Cesium3DTilesSelection {
  * {@link GltfConverters}, when the response to a network request for
  * loading the tile content was received.
  */
-struct CESIUM3DTILESSELECTION_API GltfConverterResult {
+struct CESIUM3DTILESCONTENT_API GltfConverterResult {
   /**
    * @brief The gltf model converted from a binary content. This is empty if
    * there are errors during the conversion
@@ -26,6 +26,6 @@ struct CESIUM3DTILESSELECTION_API GltfConverterResult {
    * @brief The error and warning list when converting a binary content to gltf
    * model. This is empty if there are no errors during the conversion
    */
-  ErrorList errors;
+  CesiumUtility::ErrorList errors;
 };
 } // namespace Cesium3DTilesSelection