+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/GltfConverters.h
@@ -2,7 +2,7 @@
 
 #include "Library.h"
 
-#include <Cesium3DTilesSelection/GltfConverterResult.h>
+#include <Cesium3DTilesContent/GltfConverterResult.h>
 #include <CesiumGltfReader/GltfReader.h>
 
 #include <gsl/span>