+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+### ? - ?
+
+##### Breaking Changes :mega:
+
+- Moved `ErrorList` from `Cesium3DTilesSelection` to `CesiumUtility`.
+
 ### v0.29.0 - 2023-11-01
 
 ##### Breaking Changes :mega: